Fixing the hands
As I see it, the function scale shout range between 0 and 12 and between 0 and 60. Comparing it to real clocks or just comparing degrees to hours the difference is clearly noticeable, especially the hours, since its divided in only 12 parts.

Let me know what you think..:)

Thanks for this great course btw, I'm learning a lot!